+++ changes.md
@@ -3,6 +3,7 @@
 ## Changes in dev
 1. Upgrafe three.js to r77, adjust text rendering code 
 2. Implement drawing of axes ticks on opposite side (when fTickx/y specified)  
+3. Preliminary support of candle plot (many options to be implemented)
 
 
 ## Changes in 4.5.0